requirements.txt
pandas
numpy
yfinance
ta
scikit-learn
lightgbm
joblib
matplotlib
streamlit
pyyaml
